+++ src/app/server-cluster/testing/ClusterTester.h
@@ -15,6 +15,7 @@
  */
 
 #pragma once
+#include "FabricTestFixture.h"
 #include <app/AttributeValueDecoder.h>
 #include <app/AttributeValueEncoder.h>
 #include <app/CommandHandler.h>
@@ -39,7 +40,6 @@
 #include <lib/core/TLVReader.h>
 #include <lib/support/ReadOnlyBuffer.h>
 #include <lib/support/Span.h>
-
 #include <memory>
 #include <optional>
 #include <type_traits>
@@ -47,158 +47,6 @@
 
 namespace chip {
 namespace Testing {
-
-/**
- * @brief Helper class to manage the lifecycle of a Fabric for testing.
- *
- * This encapsulates the logic required to initialize a FabricTable,
- * create a mock Fabric, and manage related certificate storage.
- * It is designed to be used by ClusterTester or a test fixture.
- */
-class FabricTestHelper
-{
-public:
-    FabricTestHelper(PersistentStorageDelegate * storage) :
-        mStorage(storage), mRootCertSpan(mRootCertDER, sizeof(mRootCertDER)), mNocSpan(mNocDER, sizeof(mNocDER))
-    {}
-
-    /**
-     * @brief Initializes the Fabric table and adds a new test fabric.
-     *
-     * @param fabricIndexOut Output parameter for the newly created FabricIndex.
-     * @return CHIP_ERROR
-     */
-
-    CHIP_ERROR SetUpFabric(FabricIndex & fabricIndexOut)
-    {
-        ReturnErrorOnFailure(mOpCertStore.Init(mStorage));
-        initParams.opCertStore         = &mOpCertStore;
-        initParams.storage             = mStorage;
-        initParams.operationalKeystore = nullptr; // Use default
-        ReturnErrorOnFailure(mfabricTable.Init(initParams));
-
-        ReturnErrorOnFailure(mfabricTable.SetFabricIndexForNextAddition(fabricIndexOut));
-        ReturnErrorOnFailure(SetUpCertificates());
-
-        CHIP_ERROR err = mfabricTable.AddNewFabricForTest(
-            mRootCertSpan, ByteSpan(), mNocSpan, ByteSpan(mSerializedOpKey.Bytes(), mSerializedOpKey.Length()), &fabricIndexOut);
-        ReturnErrorOnFailure(err);
-
-        ReturnErrorOnFailure(mfabricTable.CommitPendingFabricData());
-
-        return CHIP_NO_ERROR;
-    }
-
-    /**
-     * @brief Tears down the created fabric and cleans up storage.
-     *
-     * @param fabricIndex The FabricIndex to tear down.
-     * @return CHIP_ERROR
-     */
-    CHIP_ERROR TearDownFabric(FabricIndex & fabricIndex)
-    {
-        if (fabricIndex != chip::kUndefinedFabricIndex)
-        {
-            // Credentials::SetGroupDataProvider(nullptr); // Done in the test fixture's TearDown
-            // GroupDataProviderImpl::RemoveFabric is usually done separately
-            // However, we just need to ensure the fabric is deleted from FabricTable and storage is cleaned.
-
-            CHIP_ERROR err = CHIP_NO_ERROR;
-
-            if ((err = mOpCertStore.RemoveOpCertsForFabric(fabricIndex)) != CHIP_NO_ERROR)
-            {
-                ChipLogError(Test, "TearDownFabric: RemoveOpCertsForFabric failed: %s", err.AsString());
-            }
-            if ((err = mfabricTable.Delete(fabricIndex)) != CHIP_NO_ERROR)
-            {
-                ChipLogError(Test, "TearDownFabric: Delete fabric failed: %s", err.AsString());
-            }
-
-            mfabricTable.Shutdown();
-            mOpCertStore.Finish();
-        }
-        return CHIP_NO_ERROR;
-    }
-
-    FabricTable & GetFabricTable() { return mfabricTable; }
-
-private:
-    CHIP_ERROR SetUpCertificates()
-    {
-        Crypto::P256Keypair rootCACredentials;
-        ReturnErrorOnFailure(rootCACredentials.Initialize(Crypto::ECPKeyTarget::ECDSA));
-
-        Crypto::P256Keypair deviceOpKey;
-        ReturnErrorOnFailure(deviceOpKey.Initialize(Crypto::ECPKeyTarget::ECDSA));
-
-        ReturnErrorOnFailure(deviceOpKey.Serialize(mSerializedOpKey));
-
-        // Create temporary X.509 (DER) buffers
-        uint8_t rootCertDER_temp[chip::Credentials::kMaxDERCertLength];
-        MutableByteSpan rootCertDERSpan(rootCertDER_temp);
-
-        uint8_t nocDER_temp[chip::Credentials::kMaxDERCertLength];
-        MutableByteSpan nocDERSpan(nocDER_temp);
-
-        mRootCertSpan = MutableByteSpan(mRootCertDER);
-        mNocSpan      = MutableByteSpan(mNocDER);
-
-        chip::Credentials::X509CertRequestParams rootRequestParams;
-        rootRequestParams.SerialNumber  = kRootCertSerial;
-        rootRequestParams.ValidityStart = kCertValidityStart;
-        rootRequestParams.ValidityEnd   = kRootCertValidityDurationSeconds;
-
-        const char * rootName = "My Test Root CA";
-        ReturnErrorOnFailure(rootRequestParams.IssuerDN.AddAttribute(
-            chip::ASN1::kOID_AttributeType_CommonName, CharSpan(rootName, strlen(rootName)), true /* isPrintableString */
-            ));
-        ReturnErrorOnFailure(rootRequestParams.IssuerDN.AddAttribute(chip::ASN1::kOID_AttributeType_MatterRCACId, kTestRcacId));
-        rootRequestParams.SubjectDN = rootRequestParams.IssuerDN;
-
-        ReturnErrorOnFailure(chip::Credentials::NewRootX509Cert(rootRequestParams, rootCACredentials, rootCertDERSpan));
-
-        // Convert X.509 DER to Matter TLV
-        ReturnErrorOnFailure(chip::Credentials::ConvertX509CertToChipCert(rootCertDERSpan, mRootCertSpan));
-
-        chip::Credentials::X509CertRequestParams nocRequestParams;
-        nocRequestParams.SerialNumber  = kNocSerial;
-        nocRequestParams.ValidityStart = kCertValidityStart;
-        nocRequestParams.ValidityEnd   = kNocCertValidityDurationSeconds;
-        nocRequestParams.IssuerDN      = rootRequestParams.SubjectDN;
-        ReturnErrorOnFailure(nocRequestParams.SubjectDN.AddAttribute(chip::ASN1::kOID_AttributeType_MatterFabricId, kTestFabricId));
-        ReturnErrorOnFailure(nocRequestParams.SubjectDN.AddAttribute(chip::ASN1::kOID_AttributeType_MatterNodeId, kTestNodeId));
-        ReturnErrorOnFailure(
-            chip::Credentials::NewNodeOperationalX509Cert(nocRequestParams, deviceOpKey.Pubkey(), rootCACredentials, nocDERSpan));
-
-        ReturnErrorOnFailure(chip::Credentials::ConvertX509CertToChipCert(nocDERSpan, mNocSpan));
-
-        return CHIP_NO_ERROR;
-    }
-
-    PersistentStorageDelegate * mStorage;
-
-    // Certificates and keys
-    uint8_t mRootCertDER[chip::Credentials::kMaxDERCertLength];
-    MutableByteSpan mRootCertSpan;
-    uint8_t mNocDER[chip::Credentials::kMaxDERCertLength];
-    MutableByteSpan mNocSpan;
-    Crypto::P256SerializedKeypair mSerializedOpKey;
-
-    // Fabric-related storage
-    chip::Credentials::PersistentStorageOpCertStore mOpCertStore;
-    FabricTable mfabricTable;
-    FabricTable::InitParams initParams;
-
-    // Test constants (copied from test fixture)
-    static constexpr FabricId kTestFabricId                    = 0xDEADBEEF00000001;
-    static constexpr NodeId kTestNodeId                        = 0xDEADBEEF00000002;
-    static constexpr uint64_t kTestRcacId                      = 0x1111222233334444;
-    static constexpr uint64_t kRootCertSerial                  = 1;
-    static constexpr uint64_t kNocSerial                       = 2;
-    static constexpr uint64_t kCertValidityStart               = 0;
-    static constexpr uint32_t kRootCertValidityDurationSeconds = 315360000; // 10 years
-    static constexpr uint32_t kNocCertValidityDurationSeconds  = 31536000;  // 1 year
-};
 // Helper class for testing clusters.
 //
 // This class ensures that data read by attribute is referencing valid memory for all
@@ -229,8 +77,11 @@ private:
 class ClusterTester
 {
 public:
-    ClusterTester(app::ServerClusterInterface & cluster) :
-        mCluster(cluster), mFabricHelper(&mTestServerClusterContext.StorageDelegate())
+    ClusterTester(app::ServerClusterInterface & cluster) : mCluster(cluster), mFabricHelper(nullptr) {}
+
+    // Constructor with FabricHelper
+    ClusterTester(app::ServerClusterInterface & cluster, FabricTestFixture * fabricHelper) :
+        mCluster(cluster), mFabricHelper(fabricHelper)
     {}
 
     app::ServerClusterContext & GetServerClusterContext() { return mTestServerClusterContext.Get(); }
@@ -430,7 +281,7 @@ public:
 
     void SetFabricIndex(FabricIndex fabricIndex) { mHandler.SetFabricIndex(fabricIndex); }
 
-    FabricTestHelper & GetFabricHelper() { return mFabricHelper; }
+    FabricTestFixture * GetFabricHelper() { return mFabricHelper; }
 
 private:
     bool VerifyClusterPathsValid()
@@ -457,7 +308,7 @@ private:
     uint8_t mTlvBuffer[kTlvBufferSize];
     std::vector<std::unique_ptr<ReadOperation>> mReadOperations;
 
-    FabricTestHelper mFabricHelper;
+    FabricTestFixture * mFabricHelper;
 };
 
 } // namespace Testing
